server cleanup

+++ code/engine.vc2008/xrGame/alife_simulator.cpp
@@ -23,14 +23,25 @@
 #	include "moving_objects.h"
 #endif // DEBUG
 
-void destroy_lua_wpn_params();
 const char* alife_section = "alife";
+extern void destroy_lua_wpn_params();
 
 CALifeSimulator::CALifeSimulator(xrServer *server, shared_str *command_line) : 
 	CALifeUpdateManager(server, alife_section), CALifeSimulatorBase(server, alife_section)
 {
-	destroy_lua_wpn_params();
-	ai().set_alife(this);
+    if (strstr(Core.Params, "-keep_lua"))
+    {
+        destroy_lua_wpn_params();
+        MainMenu()->DestroyInternal(true);
+        xr_delete(g_object_factory);
+        ai().script_engine().init();
+
+#ifdef DEBUG
+        ai().moving_objects().clear();
+#endif // DEBUG
+    }
+
+	ai().set_alife				(this);
 
 	typedef IGame_Persistent::params params;
 	params &p = g_pGamePersistent->m_game_params;
